user: Set a default value for all searchNs* user preferences
Without the default the preference is never deleted from the database,
even it was disabled by the user.

This can be tested by check "all" on Special:Search and save it for
future searches. After checking "none" and selecting the main namespace
and save it for future searches there still the other searchNS options
on the database.

Bug: T291748
Change-Id: Idb3a8774b82f5f27fbeda4537dfdacf59768e8c0