watcheditem: Use lazy push for WatchlistExpiryJob
WatchlistExpiryJob is being queued in some edits and they are being pushed directly.
This means in some edits, the appserver opens a request, communicate to
eventgate and then closes it, adding around 40ms to the time to save the edit.

You can see it in flamegraph of EditAction:
https://performance.wikimedia.org/arclamp/svgs/daily/2021-09-30.excimer-wall.all.fn-EditAction.svgz

It's not an urgent job and can be pushed later through deferred update
instead. The important thing is that it gets batched through all other
jobs being pushed.

Originally added in Ica8ab92837c38.

Bug: T292048
Bug: T255502
Change-Id: I2b0612052d6b34545f50d9cd9a28da50f14a16b4