Improve side-locking logic for diffs
In case of left clicks, update the locked side correctly if the click is
inside the diff table. Only unset the side if the diff is not inside the
table.
Also, when changing side, clear any selection. This avoids some weird
behaviour when switching side (due to the existing selection on the
other side). Doing this conditionally is necessary to avoid T288421.
Also create helpers for getting/setting the current locked side, and
getting the side a Node belongs to.

Bug: T292207
Change-Id: I0e953d5336253a8b1c216d3cfbd0298d53d97842