mediawiki.base: Introduce mw.log.makeDeprecated()
Expose the main logic of `mw.log.deprecate` for tracking and logging
by itself, to allow deprecation of features that are not entire
methods or classes.

For example, to deprecate a particular parameter or other aspect
of a feature. In that case, you can't install a property getter
with mw.log.deprecate() but rather can let the developer take
responsibility for triggering the deprecation as-needed.

Usage example in the documentation.

Bug: T292489
Change-Id: I78cc15dc51bdba7d1b1f9083586170cf4808abe4