namespaceDupes.php: Handle invalid titles
Apply prefix and suffix to invalid titles just like it's applied
to conflicting titles. (Typically this will be needed when the
Foo namespace is created and there is a Foo: article already.)

Also remove a broken for loop. Presumably the intent was to
repeatedly add prefix/suffix on conflict, but actually it would
just get into an infinite loop. The behavior doesn't seem that
useful anyway.

Bug: T293407
Change-Id: Ieef6698b461e7db003dc326e5cd82accd2909540