resourceloader: Remove internal debug check in makeLoaderImplementScript()
This private method checked for debug mode, and generated the wrapping
code in either spacious form or condensed form.

The important thing to note is that $scripts being wrapped here are
always in original form, and thus the returned string still needs to
be minified before we transfer it to the client.

The only caller is makeModuleResponse(), which passes the result
to filter() for minification as-needed. Hence we don't need the extra
complication here of manually minifying it for production, we can let
the minifier take care of that instead.

Bug: T32956
Change-Id: I80559c0c257bf5c08c015373d845a6c992dc9e0b