DeleteAction: properly handle the "page not found" case
If the page cannot be found, e.g. because it was just deleted in another
request, WikiPage::doDeleteArticleReal returns an OK status with a
non-fatal "cannotdelete" message. The code in DeleteAction wasn't
accounting for that; instead, it always assumed that an OK-but-not-good
status meant the page was scheduled for deletion.

So show a success message only if the status is good or the page was
scheduled for deletion, and instead show an error if it couldn't be
deleted.

Change-Id: I7b37f96170e38a48099be5b43c5e7903b50b48e0