filerepo: Increase ttl for foreign metadata to 4h and make configurable
This is generally a bottleneck for ForeignAPIRepo. When you edit a
page with foreign images, for each image, the metadata has to be
fetched from the API (via HTTPS), including for high-DPI variants,
in serial, during the parse.

The RTT from the TCP/TLS handshakes alone can be problematic for a
page with many images.

In the event a new version of the image is uploaded, the image might
not be displayed properly until the metadata falls out of cache. It
looks like there is some adaptive TTL code, but I'm not sure it works,
since the API (e.g. ApiQueryImageInfo) does not set a Last-Modified
response header. Indeed, apart from 1 extension, nothing anywhere
sets ApiBase::getConditionalRequestData.

Change the default cache expiry from 1 hour to 4 hours, and allow
site administrators to configure this via a new 'apiMetadataExpiry'
option that can be set in $wgForeignFileRepos entries.

Bug: T261375
Change-Id: I2ff4085bd8a760387219977624d04273ada4c521