mediawiki.page.ready: Fire hook 'wikipage.indicators' with children
The parameter of the hook 'wikipage.indicators' in mediawiki.page.ready
and in mediawiki.action.edit.preview was inconsistent:

* mediawiki.page.ready fired the hook with the root element.
* mediawiki.action.edit.preview fired the hook with the children
  elements.

This should be consistent.

mediawiki.action.edit.preview fires the hook before adding the elements
to the DOM. Therefor it can't fired with the root element before adding
to DOM. To be consistent the behavior in mediawiki.page.ready must
changed to the behavior in mediawiki.action.edit.preview.

Also remove the text nodes with the newline from the value in
mediawiki.action.edit.preview to be consistent with the value in
mediawiki.page.ready.

This change is a follow-up to bea70565d96a88948b99fa657657b69f3c6df97a.

Bug: T293564
Change-Id: Ie8890e6396db071975794655c4e4ec3616891f96
(cherry picked from commit 8bec93a573f9a6a92a721a9242db8dffb88f405e)