Support variants of page content language in live preview
Before this change variants in live preview are only supported if the
content language of the Wiki project is a language with variants.

This change allows to preview variants in live preview also if the page
content language for a specific page in the database is a language with
variants.

Bug: T294052
Change-Id: I204171837072a347a75837ea63991125f2891a7b