Allow editors to customize the main page h1 heading text
Many wikis hide the main page title by default, or need to change
it to something other than the title of the page (e.g. if the page
is Main Page you can't change it to "Welcome to Wikipedia!"

They often want to do this without changing the <title> inside the
<head> of the document.

To allow them to do this, it should be possible to customize a message.

It should also be possible to show a different heading for logged in users
to meet requirements of the Minerva skin.

Bug: T293470
Bug: T255682
Change-Id: I363325653324979c8b5a06fb117d21475fb7b164