Log headers in MultiHttpClient
These are useful when debugging requests that vary behavior based on
the headers specified, like Swift's x-copy-from. Additionally,
knowing the content-length is helpful when looking at how long
the request took.

We do know that some headers are likely to contain sensitive
passwords/tokens, so try to redact those ahead of time using a regex of:

 (authorization|auth|password|cookie)

This will catch the standard Authorization header, plus Swift's
X-Auth-Token and X-Auth-Key. The regex is a best effort, but not
intended to be fully comprehensive as there's already plenty of
private data being included in logs.

Change-Id: I6e695f357f92ce8d40b54ea39002fb7858f5596e