Fix output encoding of language converter display title
The change eed3121a8fca38bbd523c7068087e611f2606bb5 added
unintentionally an HTML encoding to the language converter display
title with the syntax
  -{T|lang:Foo <i>italic</i> Bar}-
which shows raw HTML in the title. This change restores the previous
behavior for the language converter display title and applies the HTML
encoding only to the converted normal title.

This change also ensures that & is encoded to &amp; and that HTML
entities are supported like in the {{DISPLAYTITLE:}}.

Bug: T294612
Change-Id: Idc55e5756513808907be5ea95855d964a2fa704c