SpecialContributions: Let tagfilter support multiple tags
The deeper level (ContribsPager and ChangeTags::modifyDisplayQuery)
is already multi-tag aware. It doesn't make much sense to let
SpecialContributions only query via a single tag.

This feature will be useful in GrowthExperiments, which uses multiple
tags to tell multiple kinds of a question apart (like the source
of the question). We want to consturct a link to all questions asked,
not just questions from one source. See the task for details.

Bug: T294230
Change-Id: I033b5d3102906bf2e2f624e35e0c2d202f6da94f