resourceloader: Fix null type docs for FileModule::getPackageFiles
The null check for the class property packageFiles needs change to
check the return value of the function instead, because Phan would
report an issue that a null value is used in the foreach.

Phan does not understand the relation between the return value and the
class property.
When using a local variable then Phan can check
that the value is never null when used.

Change-Id: I48ac8a8fe55ae7b4e13bc94d2f2d2ede01fbc68b