parser: Make ParserOptions::registerWatcher() public and ::optionUsed() private
Follows-up fe14fdefa7 (r79018), which introduced the registerWatcher()
method with the comment that it was for internal use by the Parser.

Howevever it lacked an `@internal` annotation, and it is de-facto used
in various places now, so perhaps we should just consider it public.

This can be changed of course, but that should probably be done through
its own explicit deprecation given it's been around so long and used
in various places.

The ::optionUsed() option is only used internally, introduced by the
same commit, and has no known caller or theoreretical need to be called
outside this class. Make this private going forward.

Change-Id: Iab006be742b7bdf55a0d79509e8b827e109e9ceb