Introducing SettingsBuilder for loading settings files.
The goal is to use SettingsBuilder to load settings from JSON files
instead of relying on PHP files that manipulate global variables.

To allow for a smooth transition to the new system, config settings read
from JSON files will be applied to global variables, and it will be
possible to load JSON files programmatically from inside traditional
settings files.

Bug: T294740
Bug: T295500
Bug: T294741
Change-Id: Ibf52c660715fd0e6e67fea5169811ece9ed67cf7