Increase memory limit for DjVu metadata
With recent changes that moves reading the file to Shellbox, it seems
the current memory limit (~300KB) is extremely small and it broke
rendering of basically any DjVu file on the beta cluster, even when Shellbox is
is disabled there.

Until T292322 is resolved, this should stay. If we decide that moving
the file along the wire is better than looking them up in swift, this
should stay forever.

Bug: T275268
Change-Id: I292ee4f0ae79e7bd8afe9578c2ca0c4410c681ca