resourceloader: Make '@' work in module names
'@' is not a prohibited character in module names according to
ResourceLoader::isValidModuleName(), but attempting to use '@' in a
module name breaks because '@' is used to delimit the module name from
the version by mw.loader.implement(). Fix this by looking at the
last '@' instead of the first '@'.

This will allow us to have external libraries use the same module name
in ResourceLoader as they do in NPM; many libraries in NPM have a name
of the form '@foo/bar'.

Change-Id: I550d900e0b78e6bbc7907ab3869328f7fbd2b7ac