vue: Rename vue-composition-api module to @vue/composition-api
This way the RL module name matches the NPM module name, and code from
NPM that expects to be able to call require( '@vue/composition-api' )
will work.

Backwards compatibility is not needed because codesearch shows zero
instances of this module being used. Its only usage is in a patch in
MediaSearch that hasn't been merged yet.

Bug: T284707
Change-Id: I457dd97f00d3cbf1edb57cf61cf80afd57a1e2a7