Add new watchlist hook to better support syncing between multiple watchstars
As part of the web team's desktop improvement work there is an additional case where watchlinks need to sync state. Because the existing watchstar function is intended to bind watchstars of arbitrary page titles, we cannot use the existing callback param. Instead this approach adds a new hook that will fire on any watchstar update with an additional page title param.

Bug: T294759
Change-Id: I1d28e6046fde2fea36b2cfa1e6acd531e13270d9